% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\docType{class}
\name{cset-class}
\alias{cset-class}
\title{The CSet class}
\description{
Creates a `cset` object that represents a control set of TF-gene interactions. It inherits from the `set` class.
It is composed of 4 slots: a 2-column character dataframe named `ris` (regulatory interactions) and
a character vector named `tfs` (transcription factors) inherited from the Set class,
a character attribute `type` and a character attribute `id`.
}
\details{
#' @title The Set class
#' @description Creates a `set` object that represents a set of TF-gene interactions.
#' It is composed of 2 slots: a 2-column character dataframe named `ris` (regulatory interactions) and
#' a character vector named `tfs` (transcription factors).
#'
#' @slot ris A 2-column character dataframe of TF-gene interactions with colnames = c("tf.symbol", "gene.symbol")
#' @slot tfs A character vector of TF names
#' @import dplyr
#' @export

setClass(
  "set",
  slots = list(
    ris = "data.frame",
    tfs = "character"
  )
)

# setValidity("set", function(object) {
#   ## TODO
#   ## tfs have to be ordered and unique
# })

#' @title Constructor function of a set object.
#' @description Constructor function of a set object.
#' @param ris A 2-column character dataframe of TF-gene interactions.
#' @param tfs A character vector of TF names.
#' @return A `set` object.
#' @export

set <-  function(ris, tfs) {
  stopifnot(is.data.frame(ris))
  stopifnot(is.character(tfs))
  ris %>% dplyr::distinct() ## maybe issue a warning in case tfs and or ris are not unique?
  tfs <- unique(sort(tfs))

  new("set", ris = ris, tfs = tfs)
}
}
\section{Slots}{

\describe{
\item{\code{ris}}{A 2-column character dataframe of TF-gene interactions with colnames = c("tf.symbol", "gene.symbol")}

\item{\code{tfs}}{A character vector of TF names}

\item{\code{type}}{A character in c("positive", "negative")}

\item{\code{id}}{A character name}
}}

