% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\name{cset}
\alias{cset}
\title{Constructor function of a cset object.}
\usage{
cset(set, type, id)
}
\arguments{
\item{set}{A `set` object}

\item{type}{A character in c("positive", "negative")}

\item{id}{A character name}
}
\value{
A `cset` object.
}
\description{
Constructor function of a set object.
}
