% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-class.R
\docType{class}
\name{evalset-class}
\alias{evalset-class}
\title{The EvalSet class}
\description{
Performs the evaluation of a given input set against selected controls sets (1 positive and 1 negative).
It is thus composed of 3 input `set` objects.
}
\section{Slots}{

\describe{
\item{\code{pos_set}}{A cset object with type = "positive".}

\item{\code{neg_set}}{A cset object with type = "negative".}

\item{\code{pred_set}}{A pset object containing predictions to be evaluated against controls.}
}}

