% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\docType{methods}
\name{true_pos}
\alias{true_pos}
\alias{true_pos,evalset-method}
\alias{false_pos}
\alias{false_pos,evalset-method}
\alias{true_neg}
\alias{true_neg,evalset-method}
\alias{false_neg}
\alias{false_neg,evalset-method}
\alias{pred_pos}
\alias{pred_pos,evalset-method}
\alias{pred_neg}
\alias{pred_neg,evalset-method}
\alias{actual_pos}
\alias{actual_pos,evalset-method}
\alias{actual_neg}
\alias{actual_neg,evalset-method}
\alias{total_pop}
\alias{total_pop,evalset-method}
\alias{sensitivity}
\alias{sensitivity,evalset-method}
\alias{specificity}
\alias{specificity,evalset-method}
\alias{precision}
\alias{precision,evalset-method}
\alias{fdr}
\alias{fdr,evalset-method}
\alias{outer_set}
\alias{outer_set,evalset-method}
\alias{inner_set}
\alias{shuffle_scores}
\alias{shuffle_scores,evalset-method}
\title{Get TP number}
\usage{
true_pos(x)

false_pos(x)

true_neg(x)

false_neg(x)

pred_pos(x)

pred_neg(x)

actual_pos(x)

actual_neg(x)

total_pop(x)

sensitivity(x)

specificity(x)

precision(x)

fdr(x)

outer_set(x)

inner_set(x)

shuffle_scores(x)
}
\arguments{
\item{x}{An `evalset` object}
}
\value{
An integer

An integer

An integer

An integer

An integer

An integer

An integer

An integer

An integer

A numeric

A numeric

An numeric

An numeric

A `set` object

A `set` object

An `evalset` object
}
\description{
Get the set of RIs in the prediction set that are not in either control set,
and thus not taken into account for classic statistics computation

Get the set of RIs in the prediction set that are not in either control set,
and thus not taken into account for classic statistics computation

Shuffle scores of predicted set, each score column independently
}
