% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-class.R
\name{evalset}
\alias{evalset}
\title{Constructor function of the EvalSet object}
\usage{
evalset(pos_set, neg_set, pred_set, out_tfs, out_ris)
}
\arguments{
\item{pos_set}{A set object to be used a positive control.}

\item{neg_set}{A set object to be used a negative control.}

\item{pred_set}{A set object containing predictions to be evaluated against controls.}
}
\value{
An `evalset` object.
}
\description{
Constructor function of an EvalSet object in order to perform the evaluation of a predicted set against control sets.
}
