% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_inputs.R
\name{format_cset}
\alias{format_cset}
\title{Generate a cset from the input parameters.}
\usage{
format_cset(id, type, dir = "control_sets")
}
\arguments{
\item{id}{The path to a prediction file.}

\item{type}{Type from c("negative", "positive")}

\item{dir}{Optional, by default internal directory "control_sets"}
}
\description{
Reads the control file and builds the object.
}
\author{
Claire Rioualen
}
