% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_inputs.R
\name{format_pset}
\alias{format_pset}
\title{Generate a pset from the input file.}
\usage{
format_pset(file)
}
\arguments{
\item{file}{The input prediction file, a 2+ column file with colnames: c("tf_bnum", "gene_bnum", ...)}
}
\description{
Format the input prediction set and builds the object.
}
\author{
Claire Rioualen
}
