% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{generate_confusion_matrix}
\alias{generate_confusion_matrix}
\title{Generate a confusion matrix for a given evalset object.}
\usage{
generate_confusion_matrix(x)
}
\arguments{
\item{x}{An `evalset` object.}
}
\value{
A dataframe.
}
\description{
Generate a confusion matrix for a given evalset object, composed of a `pset` to be evaluated and a positive and a negative
}
\author{
Claire Rioualen
}
