% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{generate_pr_curve}
\alias{generate_pr_curve}
\title{Generate a precision-recall curve for a given evalset object.}
\usage{
generate_pr_curve(x, s)
}
\arguments{
\item{evalset}{An evalset object.}
}
\value{
A plot.
}
\description{
Generate a precision-recall curve for a given evalset object.
}
\author{
Claire Rioualen
}
