% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{generate_roc_curve}
\alias{generate_roc_curve}
\title{Generate a ROC curve for a given evalset object.}
\usage{
generate_roc_curve(x)
}
\arguments{
\item{evalset}{An evalset object.}

\item{score}{An character vector indicating which score to process.}
}
\value{
A list containing a plot and a dataframe
}
\description{
Generate a ROC curve for a given evalset object..
}
\author{
Claire Rioualen
}
