% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{generate_venn_diagram}
\alias{generate_venn_diagram}
\title{Generate a Venn diagram using the eulerr package.}
\usage{
generate_venn_diagram(evalset, style = 1, universe = FALSE)
}
\arguments{
\item{evalset}{A list of Set objects.}

\item{style}{An option in c(1, 2, 3). Default to 1 (classic venn plot)}

\item{universe}{A logical indicating wether to plot universe of interactions. Defaults to FALSE.}
}
\value{
A plot.
}
\description{
Generate a Venn diagram given a list of Set objects.
}
\author{
Claire Rioualen
}
