% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data.R
\name{get_ensembl_genes}
\alias{get_ensembl_genes}
\title{Get genes from Ensembl}
\usage{
get_ensembl_genes()
}
\value{
A character vector.
}
\description{
Get genes from Ensembl. Note: I custom-made the currently available file. Should be inddcluded in future makefile.
}
\author{
Claire Rioualen
}
