% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data.R
\name{get_regulondb_genes}
\alias{get_regulondb_genes}
\title{Get genes from RegulonDB}
\usage{
get_regulondb_genes()
}
\value{
A character vector.
}
\description{
Get genes from RegulonDB. Note: the RegulonDB files have to be previously downloaded (to be included in setup - maybe a makefile?).
}
\author{
Claire Rioualen
}
