% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data.R
\name{get_regulondb_tfs}
\alias{get_regulondb_tfs}
\title{Get transcription factors from RegulonDB}
\usage{
get_regulondb_tfs()
}
\arguments{
\item{file}{A TFSet file downloaded from RegulonDB.}
}
\value{
A character vector.
}
\description{
Get list of transcription factor coding gene names from RegulonDB
}
\details{
Basic funtions to get and work with lists of genes and tfs.
This section should evolve once we have a proper standard: right now there are inconsistencies between RegulonDB and Ensembl gene annotations.
Eventually we want to have a proper function to filter/cconvert gene ids and names.

We should also be able to convert tf names properly between RegulonDB and Zika, while giving particualr attention to dimeric names.
}
\author{
Claire Rioualen
}
