% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{get_universe}
\alias{get_universe}
\title{Get universe of regulatory interactions}
\usage{
get_universe(x)
}
\arguments{
\item{evalset}{An `evalset` object.}
}
\value{
universe A `set` object of all combinations.
}
\description{
Get set of all possible RIs given the evalset list of TFs and all genes in RegulonDB (!= genes in ensembl)
}
\author{
Claire Rioualen
}
