% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\docType{class}
\name{pset-class}
\alias{pset-class}
\title{The PSet class}
\description{
Creates a `pset` object that represents a prediccted set of TF-gene interactions. It inherits from the `set` class.
It is composed of 3 slots: a 2-column character dataframe named `ris` (regulatory interactions) and
a character vector named `tfs` (transcription factors) inherited from the Set class,
and a 1-column dataframe containing scores associated with each regulatory interaction.
}
\section{Slots}{

\describe{
\item{\code{ris}}{A 2-column character dataframe of TF-gene interactions with colnames = c("tf.symbol", "gene.symbol")}

\item{\code{tfs}}{A character vector of TF names}

\item{\code{scores}}{A dataframe of 1 column with a number of rows equal to that of set@ris.}
}}

