% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\name{pset}
\alias{pset}
\title{Constructor function of a pset object.}
\usage{
pset(set, scores, id)
}
\arguments{
\item{set}{A `set` object.}

\item{scores}{A dataframe of 1 column with a number of rows equal to that of set@ris.}
}
\value{
A `pset` object.
}
\description{
Constructor function of a set object.
}
