% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eval.R
\name{run_eval}
\alias{run_eval}
\title{Run the evaluation of a prediction set.}
\usage{
run_eval(testfile, neg_set_id = "all_negative", pos_set_id = "all_positive")
}
\arguments{
\item{testfile}{The path to a prediction file.}

\item{neg_set_id}{Optional, the ID of the negative set to be used from c("all_negative", "test_neg")}

\item{pos_set_id}{Optional, the ID of the positive set to be used from c("all_positive", "test_pos")}
}
\description{
Format the input prediction set, filtters it, and generates an evaluation report.
}
\author{
Claire Rioualen
}
