% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_inputs.R
\name{select_tfs}
\alias{select_tfs}
\title{Filter the 3 sets' TFs before evaluation}
\usage{
select_tfs(neg_set, pos_set, pred_set)
}
\arguments{
\item{neg_set}{A cset object of type "negative"}

\item{pos_set}{A cset objecct of type "positive"}

\item{pred_set}{A pset object}
}
\description{
Get the list of common TFs between all 3 sets and select RIs accordingly
}
\author{
Claire Rioualen
}
