% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\docType{class}
\name{set-class}
\alias{set-class}
\title{The Set class}
\description{
The Set class
}
\section{Slots}{

\describe{
\item{\code{ris}}{A 2-column character dataframe}

\item{\code{tfs}}{A character vector of TF names}
}}

