% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-class.R
\name{set}
\alias{set}
\title{Constructor function of a set object.}
\usage{
set(ris, tfs)
}
\arguments{
\item{ris}{A 2-column character dataframe of TF-gene interactions.}

\item{tfs}{A character vector of TF names.}
}
\value{
A `set` object.
}
\description{
Constructor function of a set object.
}
