% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-methods.R
\name{summarize}
\alias{summarize}
\title{Generate a dataframe of sets properties.}
\usage{
summarize(x)
}
\arguments{
\item{set}{A Set object.}
}
\value{
A dataframe.
}
\description{
Generate a dataframe of sets properties.
}
\author{
Claire Rioualen
}
