% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Generate a table of stats for a given evalset object.}
\usage{
summarize_stats(x)
}
\arguments{
\item{evalset}{An `evalset` object.}
}
\value{
A dataframe.
}
\description{
Generate a table of stats for a given evalset object: sensitivity, specificity, etc.
}
\author{
Claire Rioualen
}
