% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_evalset-methods.R
\name{write_report}
\alias{write_report}
\title{Generate a report of the analyses.}
\usage{
write_report(x)
}
\arguments{
\item{eval_list}{An list of Evaluation objects.}
}
\value{
A plot.
}
\description{
Generate a report summarizing all the statistics en graphes generated for one or several evaluations.
}
\author{
Claire Rioualen
}
