% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_data.R
\name{get_regulondb_tu_first_gene}
\alias{get_regulondb_tu_first_gene}
\title{Get first genes of TUs from RegulonDB}
\usage{
get_regulondb_tu_first_gene()
}
\value{
A character vector.
}
\description{
Get TU first genes from RegulonDB. Note: the RegulonDB files have to be previously downloaded (to be included in setup - maybe a makefile?).
}
\author{
Claire Rioualen
}
