% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_set-methods.R, R/02_evalset-methods.R
\name{summarize}
\alias{summarize}
\title{Generate a dataframe of sets properties.}
\usage{
summarize(x)

summarize(x)
}
\arguments{
\item{set}{A Set object.}

\item{evalset}{An `evalset` object.}
}
\value{
A dataframe.

A dataframe.
}
\description{
Generate a dataframe of sets properties.

Generate a table of stats for a given evalset object: sensitivity, specificity, etc.
}
\author{
Claire Rioualen

Claire Rioualen
}
